Added TrackObject prefab
Test scene: Interactive Elements Test 2

This uses Splines, and the built-in spline animate script therein. Tracked objects use the InteractiveElementVelocityAnimate script to animate the ElapsedTime variable in the Spline Animate script.

To animate a different type of object, just edit the mesh and collider on the default Track Object prefab.