More audio hookups: Oven ambience, cook/burn
Oven ambience:
This is a simple empty prefab with an audio source that I lazily attached to the Slime prefab so that it gets put in every level. This will play on instantiate, and is set to loop.

Cook/Burn: These are audio sources attached to the PARTICLE EFFECTS for burning/cooking: BurnedParticles and CookedParticles. They play when instantiated.